+++ src/Symfony/Component/Security/Http/Firewall.php
@@ -108,10 +108,7 @@ class Firewall implements EventSubscriberInterface
         }
     }
 
-    /**
-     * @return array
-     */
-    public static function getSubscribedEvents()
+    public static function getSubscribedEvents(): array
     {
         return [
             KernelEvents::REQUEST => ['onKernelRequest', 8],